requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
bcrypt==4.2.0
blinker==1.8.2
Brotli==1.1.0
cachetools==4.2.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
cryptography==43.0.3
cssselect2==0.8.0
DateTime==5.5
debugpy==1.8.8
decorator==5.1.1
Django==5.1.4
django-cors-headers==4.7.0
django-jazzmin==3.0.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.58.1
future==0.18.2
greenlet==3.1.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.4.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
numpy==2.0.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.1
pandas==2.2.3
paramiko==3.5.0
parso==0.8.4
pefile==2021.5.24
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydyf==0.11.0
Pygments==2.18.0
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pypdf==5.5.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
reportlab==4.4.1
requests==2.32.3
setuptools==75.3.0
six==1.16.0
spyder-kernels==2.5.2
SQLAlchemy==2.0.35
sqlparse==0.5.3
stack-data==0.6.3
style==1.1.0
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
unicorn==2.1.3
update==0.0.1
uritools==5.0.0
urllib3==2.2.3
waitress==3.0.2
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.0.4
WTForms==3.2.1
xhtml2pdf==0.2.17
zipp==3.20.2
zope.interface==7.1.1
zopfli==0.2.3.post1
docopt==0.6.2
environ==1.0
num2words==0.5.14
qrcode==8.2
